# Conflicts:
#	pom.xml
#	src/main/java/com/example/skyfast_2_0/entity/Airline.java
#	src/main/java/com/example/skyfast_2_0/entity/Airport.java
#	src/main/java/com/example/skyfast_2_0/entity/Baggage.java
#	src/main/java/com/example/skyfast_2_0/entity/Booking.java
#	src/main/java/com/example/skyfast_2_0/entity/Feedback.java
#	src/main/java/com/example/skyfast_2_0/entity/Maintenance.java
#	src/main/java/com/example/skyfast_2_0/entity/News.java
#	src/main/java/com/example/skyfast_2_0/entity/Passenger.java
#	src/main/java/com/example/skyfast_2_0/entity/Payment.java
#	src/main/java/com/example/skyfast_2_0/entity/Paymenttype.java
#	src/main/java/com/example/skyfast_2_0/entity/Refund.java
#	src/main/java/com/example/skyfast_2_0/entity/Route.java
#	src/main/java/com/example/skyfast_2_0/entity/Seat.java
#	src/main/java/com/example/skyfast_2_0/entity/TicketBaggage.java
#	src/main/java/com/example/skyfast_2_0/entity/TicketBaggageId.java
#	src/main/java/com/example/skyfast_2_0/entity/User.java
#	src/main/java/com/example/skyfast_2_0/repository/K_BookingRepository.java
#	src/main/resources/application.properties
#	src/main/resources/templates/Dashboard.html
#	src/main/resources/templates/SkyFast_2_0.html
#	src/main/resources/templates/UserDetail.html
#	src/main/resources/templates/userlist.html

+++ src/main/java/com/example/skyfast_2_0/entity/Airplane.java
@@ -11,6 +11,7 @@ import lombok.Setter;
 @Entity
 @Table(name = "airplane")
 public class Airplane {
+
     @Id
     @Column(name = "id", nullable = false)
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -37,15 +38,15 @@ public class Airplane {
 
     @NotNull
     @Column(name = "total_length", nullable = false)
-    private Float totalLength;
+    private Integer totalLength;
 
     @NotNull
     @Column(name = "wingspan", nullable = false)
-    private Float wingspan;
+    private Integer wingspan;
 
     @NotNull
     @Column(name = "height", nullable = false)
-    private Float height;
+    private Integer height;
 
     @Size(max = 255)
     @NotNull
@@ -65,5 +66,4 @@ public class Airplane {
     @ManyToOne(fetch = FetchType.LAZY, optional = false)
     @JoinColumn(name = "airline_id", nullable = false)
     private Airline airline;
-
 }
